project/requirements.txt
-i https://pypi.org/simple
asgiref==3.6.0 ; python_version >= '3.7'
cachetools==5.3.0 ; python_version ~= '3.7'
certifi==2022.12.7 ; python_version >= '3.6'
charset-normalizer==3.0.1
django==4.1.7
google-ads==20.0.0
google-api-core==2.11.0 ; python_version >= '3.7'
google-auth==2.16.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
google-auth-oauthlib==0.8.0 ; python_version >= '3.6'
google-cloud-bigquery==3.6.0
google-cloud-core==2.3.2 ; python_version >= '3.7'
google-cloud-storage==2.7.0
google-crc32c==1.5.0 ; python_version >= '3.7'
google-resumable-media==2.4.1 ; python_version >= '3.7'
googleapis-common-protos==1.58.0 ; python_version >= '3.7'
grpcio==1.51.3 ; python_version >= '3.7'
grpcio-status==1.51.3 ; python_version >= '3.6'
idna==3.4 ; python_version >= '3.5'
oauthlib==3.2.2 ; python_version >= '3.6'
packaging==23.0 ; python_version >= '3.7'
proto-plus==1.22.2 ; python_version >= '3.6'
protobuf==4.22.0 ; python_version >= '3.7'
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyyaml==6.0 ; python_version >= '3.6'
requests==2.28.2 ; python_version >= '3.7' and python_version < '4'
requests-oauthlib==1.3.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
rsa==4.9 ; python_version >= '3.6'
setuptools==67.4.0 ; python_version >= '3.7'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.3 ; python_version >= '3.5'
urllib3==1.26.14 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
